auditlog: Add ResourceContext
