+++ setup.cfg
@@ -35,7 +35,7 @@ tests =
     pytest-black-ng>=0.4.0
     pytest-invenio>=3.0.0,<4.0.0
     invenio-app>=2.0.0,<3.0.0
-    invenio-users-resources>=7.0.0,<8.0.0 # Only for tests
+    invenio-users-resources>=7.0.0,<8.0.0
     Sphinx>=4.5.0
     invenio-db[mysql,postgresql]>=2.0.0,<3.0.0
 
@@ -54,11 +54,11 @@ invenio_db.models =
 invenio_celery.tasks =
     invenio_drafts_resources = invenio_drafts_resources.services.records.tasks
 invenio_audit_logs.actions =
-    record.publish = invenio_drafts_resources.auditlog_actions:RecordPublishAuditLog
-    draft.create = invenio_drafts_resources.auditlog_actions:DraftCreateAuditLog
-    draft.edit = invenio_drafts_resources.auditlog_actions:DraftEditAuditLog
-    draft.delete = invenio_drafts_resources.auditlog_actions:DraftDeleteAuditLog
-    draft.new_version = invenio_drafts_resources.auditlog_actions:DraftNewVersionAuditLog
+    record.publish = invenio_drafts_resources.auditlog.actions:RecordPublishAuditLog
+    draft.create = invenio_drafts_resources.auditlog.actions:DraftCreateAuditLog
+    draft.edit = invenio_drafts_resources.auditlog.actions:DraftEditAuditLog
+    draft.delete = invenio_drafts_resources.auditlog.actions:DraftDeleteAuditLog
+    draft.new_version = invenio_drafts_resources.auditlog.actions:DraftNewVersionAuditLog
 
 [build_sphinx]
 source-dir = docs/
